almost done

+++ sources/moves.c
@@ -6,7 +6,7 @@
 /*   By: vtestut <vtestut@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/02/06 21:13:24 by vtestut           #+#    #+#             */
-/*   Updated: 2024/02/11 19:56:29 by vtestut          ###   ########.fr       */
+/*   Updated: 2024/02/11 20:19:52 by vtestut          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,8 +14,8 @@
 
 int	move_front(t_game *game)
 {
-	double	new_x;
-	double	new_y;
+	float	new_x;
+	float	new_y;
 
 	new_x = game->player.pos_x + game->player.dir_x * SPEED;
 	new_y = game->player.pos_y + game->player.dir_y * SPEED;
@@ -24,8 +24,8 @@ int	move_front(t_game *game)
 
 int	move_back(t_game *game)
 {
-	double	new_x;
-	double	new_y;
+	float	new_x;
+	float	new_y;
 
 	new_x = game->player.pos_x - game->player.dir_x * SPEED;
 	new_y = game->player.pos_y - game->player.dir_y * SPEED;
@@ -34,8 +34,8 @@ int	move_back(t_game *game)
 
 int	move_left(t_game *game)
 {
-	double	new_x;
-	double	new_y;
+	float	new_x;
+	float	new_y;
 
 	new_x = game->player.pos_x + game->player.dir_y * SPEED;
 	new_y = game->player.pos_y - game->player.dir_x * SPEED;
@@ -44,8 +44,8 @@ int	move_left(t_game *game)
 
 int	move_right(t_game *game)
 {
-	double	new_x;
-	double	new_y;
+	float	new_x;
+	float	new_y;
 
 	new_x = game->player.pos_x - game->player.dir_y * SPEED;
 	new_y = game->player.pos_y + game->player.dir_x * SPEED;
